refactor(redux): Move actionTypes to index file
